# Speed up type checking by caching argument inference context