This speeds up self check by about 1.5% on my computer, according to `perf_compare.py`:
```
...
=== Results ===

master                    4.264s (0.0%) | stdev 0.024s 
HEAD                      4.201s (-1.5%) | stdev 0.030s 
```
I noticed this bottleneck when I was looking at a CPU profile generated using the script from #19322.